Resolve circular dependency by moving funcs that require Wallet to transactions.py
