+++ ocean_lib/web3_internal/utils.py
@@ -11,7 +11,6 @@ from eth_keys import keys
 from eth_utils import big_endian_to_int, decode_hex
 from ocean_lib.enforce_typing_shim import enforce_types_shim
 from ocean_lib.web3_internal.constants import DEFAULT_NETWORK_NAME, NETWORK_NAME_MAP
-from ocean_lib.web3_internal.wallet import Wallet
 from ocean_lib.web3_internal.web3_overrides.signature import SignatureFix
 from ocean_lib.web3_internal.web3_provider import Web3Provider
 
@@ -123,20 +122,6 @@ def get_network_id() -> int:
 
 
 @enforce_types_shim
-def sign_hash(msg_hash: str, wallet: Wallet) -> str:
-    """
-    This method use `personal_sign`for signing a message. This will always prepend the
-    `\x19Ethereum Signed Message:\n32` prefix before signing.
-
-    :param msg_hash:
-    :param wallet: Wallet instance
-    :return: signature
-    """
-    s = wallet.sign(msg_hash)
-    return s.signature.hex()
-
-
-@enforce_types_shim
 def ec_recover(message: str, signed_message: str):
     """
     This method does not prepend the message with the prefix `\x19Ethereum Signed Message:\n32`.
@@ -174,47 +159,3 @@ def get_ether_balance(address: str) -> int:
 
 def from_wei(wei_value: int) -> Decimal:
     return Web3Provider.get_web3().fromWei(wei_value, "ether")
-
-
-def send_ether(from_wallet: Wallet, to_address: str, ether_amount: int):
-    w3 = Web3Provider.get_web3()
-    if not w3.isChecksumAddress(to_address):
-        to_address = w3.toChecksumAddress(to_address)
-
-    tx = {
-        "from": from_wallet.address,
-        "to": to_address,
-        "value": w3.toWei(ether_amount, "ether"),
-    }
-    _ = w3.eth.estimateGas(tx)
-    tx = {
-        "from": from_wallet.address,
-        "to": to_address,
-        "value": w3.toWei(ether_amount, "ether"),
-        "gas": 500000,
-    }
-    wallet = Wallet(w3, private_key=from_wallet.key, address=from_wallet.address)
-    raw_tx = wallet.sign_tx(tx)
-    tx_hash = w3.eth.sendRawTransaction(raw_tx)
-    receipt = w3.eth.waitForTransactionReceipt(tx_hash, timeout=30)
-    return receipt
-
-
-def cancel_or_replace_transaction(
-    from_wallet, nonce_value, gas_price=None, gas_limit=None
-):
-    w3 = Web3Provider.get_web3()
-    tx = {"from": from_wallet.address, "to": from_wallet.address, "value": 0}
-    gas = gas_limit if gas_limit is not None else w3.eth.estimateGas(tx)
-    tx = {
-        "from": from_wallet.address,
-        "to": from_wallet.address,
-        "value": 0,
-        "gas": gas + 1,
-    }
-
-    wallet = Wallet(w3, private_key=from_wallet.key, address=from_wallet.address)
-    raw_tx = wallet.sign_tx(tx, fixed_nonce=nonce_value, gas_price=gas_price)
-    tx_hash = w3.eth.sendRawTransaction(raw_tx)
-    receipt = w3.eth.waitForTransactionReceipt(tx_hash, timeout=30)
-    return receipt
